Create workspace.xml
Ok So Im starting the packaging of M_E_GA. The package it's self will be the M_E_Engine and M_E_GA_Base Which make up the core functionality of MEGA Im still working on the documentation of it and will include the templates for fitness functions and experiments as well as the Genetic Programming classes eventually... hopefully soon. Once I have everything set up I will merge the branches so everyone can work under the package.