Updating bug fixes, adding delimit delete probability
Ok So I found a bug in the logging where there are two generation log entries per generation one empty and the other containing the generation logs.

Introduced the delimit delete independent probability. This allows setting the probability od deleting a delimiter pair when mutating on a delimiter gene. this counter balance with insert delimiter pair making it easier to balance the introduction of new delimiter pairs into the organism.