Update M_E_Engine.py
Updated Parameters to reflect the decided terminology moving forward.  I realized that when I Was developing this I didn't have a fully developed perspective on how things should be names and I'm constantly referring to encodings as meta genes and its annoying and confusing to others looking in the code and not seeing Meta Genes anywhere. So now its clearer hopefully.