Implimented MetaGene deletion
Works great for the mostpart. There is a certain unknown gene configuration that is pretty rare that causes an infinite loop not sure how to fix it yet but so far not too shabby. Going to add in some features allowing for metagenome dumping and uploading too among other things.