Update M_E_Engine.py
Added Metagene deletion. Works well. cache  is used to determine the number of metagenes and genes that fall out of the LRU cache are deleted based on time in the delete basket unpacked to replace themselves in the hierarchy with their contents so their data is preserved until the higher order genes containing them are also deleted. Reduces information loss and gives higher order metagenes the opportunity to refine the data more easily before it is lost to selection pressure.