Codex/update delete metagene logic (#12)
* Mega dev (#9)

* Update M_E_GA_Base.py

Modified gene variables and params to reflect the decided naming convention. See change log for details.

* Create SP_NN_Single_Run.py

This is an updated version of the SP_NN Single Run for the SP_NN framework its here just to keep it separate from my local files while I continue to test and modify the MEGA framework.

* Update M_E_Engine.py

Updated Parameters to reflect the decided terminology moving forward.  I realized that when I Was developing this I didn't have a fully developed perspective on how things should be names and I'm constantly referring to encodings as meta genes and its annoying and confusing to others looking in the code and not seeing Meta Genes anywhere. So now its clearer hopefully.

* Update CHANGELOG.md

Updated changelog to reflect unreleased changes that will be present in 2.0.0b0

* Implimented MetaGene deletion

Works great for the mostpart. There is a certain unknown gene configuration that is pretty rare that causes an infinite loop not sure how to fix it yet but so far not too shabby. Going to add in some features allowing for metagenome dumping and uploading too among other things.

* Update M_E_Engine.py

Should fix halting problem. fingers crossed.

* test

* update

* Update M_E_Engine.py

Added Metagene deletion. Works well. cache  is used to determine the number of metagenes and genes that fall out of the LRU cache are deleted based on time in the delete basket unpacked to replace themselves in the hierarchy with their contents so their data is preserved until the higher order genes containing them are also deleted. Reduces information loss and gives higher order metagenes the opportunity to refine the data more easily before it is lost to selection pressure.

* Update M_E_Engine.py

Update to ensure sync.

* Update M_E_GA_Base.py

Update for sync

* removed excess files that snuck in.

* Update to include metagene deletion.

* Update

Update to logging structure and examples.  Visualization in the experiment runner in src. Still working on everything no official release.

* Update CHANGELOG.md

Update to include addition and changing of the logging system to make it more modular and manageable. as well as allowing for live analysis of the algorithm

* Update setup.py

* Update

* Update GA_Logger.py

* Update setup.py

* Update MANIFEST.in

* Update M_E_GA_Base.py

Fixing imports Again!! Eventually Ill get used to relative imports.

* Delete src/M_E_GA/M_E_GA_fitness_funcs.py

* Code fix

* cleaning up

* Update setup.py

Error in the logger being included. Man I hate Package building. only a couple hours in and Im already on 2.0.0b1

* Update M_E_GA_Base.py

Added Docstrings

* Update M_E_Engine.py

Update to include Docstrings

* Update GA_Logger.py

Update to include Docstrings.

* Update experiment_runner.py

Update to include Docstrings

* Update README.md

update to fix wording.

* Big SRP Refactor

* Fix tests

* Update CHANGELOG.md

* Update experiment_runner.py

* Update and rename M_E_GA_fitness_funcs.py to Leading_ones.py

* Update experiment_runner.py

update to use new logging

* Update CHANGELOG.md

* Update CHANGELOG.md

* Update setup.py

* Create __init__.py

* Create __init__.py

* Update setup.py

* Update CHANGELOG.md

* Delete dist directory

* Meta gene deletion bug (#8)

* Delete dist directory

* Update experiment_runner.py

* Update Leading_ones.py

* Update to Fix Metagene deletion bug #7

---------

Co-authored-by: Noli <nolialsea@gmail.com>

* Update setup.py

fixed duplicate line

* Delete examples/__pycache__ directory

* Create requirements.txt

* Remove deleted metagene from encodings

---------

Co-authored-by: Noli <nolialsea@gmail.com>